requirements.txt
pyglet==1.5.26
rstr==3.1.0
pyparallel==0.2.2
Pyglet-gui==0.1
pylsl==1.16.1